Move EditorWidgetActions to api directory
This is consistent with how actions are defined in other plugins.
It also prevents a circular import in the next commit.
